requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-filter==23.3
django-redis==5.4.0
celery==5.3.4
elasticsearch==8.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-spectacular==0.26.5
